requirements.txt
Flask
python-dotenv
requests
pandas
botocore
